move deletionResponseModal to modals
